requirements.txt
tensorflow==2.13.1          
keras==2.13.1               
opencv-python==4.8.1.78
mtcnn==0.1.1              
Pillow==10.0.1
numpy==1.24.3               
pandas==2.0.3
scikit-learn==1.3.2
albumentations==1.3.1
matplotlib==3.7.3
seaborn==0.12.2
streamlit==1.28.2
tqdm==4.66.1
h5py==3.9.0            
requests==2.31.0
